funcional, faltan graficas y editar distribucion

+++ pages/api/despacho/[id].js
@@ -1,7 +1,16 @@
 import { pool } from "@/config/db";
 import moment from "moment";
+import { authOptions } from 'pages/api/auth/[...nextauth]'
+import { getServerSession } from "next-auth/next"
 
 export default async function handler(req, res) {
+  const session = await getServerSession(req, res, authOptions)
+
+  if (!session || session.rol != "superadmin") {
+    res.status(403).json({ message: "Por favor, contacte al administrador" });
+    return;
+  }
+  
   switch (req.method) {
     case "DELETE":
       return await eliminarDespacho(req, res);
